requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
torch>=2.0.0
transformers>=4.38.0
tree-sitter==0.21.3
pymongo==4.6.3
redis==5.0.4
celery==5.3.6
pytest==8.1.1
pytest-asyncio==0.23.6
httpx==0.27.0
scikit-learn>=1.4.0
numpy>=1.26.0
pydantic>=2.6.0
pyyaml>=6.0.1
python-dotenv>=1.0.1
lightgbm>=4.3.0
scipy>=1.13.0
tqdm>=4.66.2
